Add styling to the homepage and the calculator page
