+++ src/App.js
@@ -1,30 +1,18 @@
-import { Route, Routes, NavLink } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import Homepage from './components/Homepage';
 import Calculator from './components/Calculator';
-import Logo from './components/Logo';
-
-const links = [
-  { path: '/', text: 'Home' },
-  { path: 'calculator', text: 'Calculator' },
-  { path: 'quote', text: 'Quote' },
-];
+import Header from './components/Header';
+import Quote from './components/Quote';
+import './App.css';
 
 function App() {
   return (
     <>
-      <Logo />
-      <nav className="navbar">
-        <ul>
-          {links.map((link) => (
-            <li key={link.text}>
-              <NavLink to={link.path}>{link.text}</NavLink>
-            </li>
-          ))}
-        </ul>
-      </nav>
+      <Header />
       <Routes>
         <Route path="/" element={<Homepage />} />
         <Route path="/calculator" element={<Calculator />} />
+        <Route path="/quote" element={<Quote />} />
       </Routes>
     </>
   );
