+++ src/components/Homepage.js
@@ -1,19 +1,23 @@
+import styles from '../styles/Homepage.module.css';
+
 const Homepage = () => (
-  <>
-    <h1>Welcome to our page!</h1>
-    <p>
-      Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-      tempor incididunt ut labore et dolore magna aliqua. Sollicitudin tempor
-      id eu nisl nunc mi ipsum faucibus. Egestas diam in arcu cursus euismod
-      quis viverra nibh cras.
-    </p>
-    <p>
-      Vel fringilla est ullamcorper eget nulla. Sit amet porttitor eget dolor
-      morbi non arcu. Elementum eu facilisis sed odio morbi quis commodo odio.
-      Non tellus orci ac auctor augue mauris. Accumsan lacus vel facilisis
-      volutpat est velit. Aliquam purus sit amet luctus.
-    </p>
-  </>
+  <div className={styles.wrapper}>
+    <article>
+      <h1>Welcome to our page!</h1>
+      <p>
+        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+        tempor incididunt ut labore et dolore magna aliqua. Sollicitudin tempor
+        id eu nisl nunc mi ipsum faucibus. Egestas diam in arcu cursus euismod
+        quis viverra nibh cras.
+      </p>
+      <p>
+        Vel fringilla est ullamcorper eget nulla. Sit amet porttitor eget dolor
+        morbi non arcu. Elementum eu facilisis sed odio morbi quis commodo odio.
+        Non tellus orci ac auctor augue mauris. Accumsan lacus vel facilisis
+        volutpat est velit. Aliquam purus sit amet luctus.
+      </p>
+    </article>
+  </div>
 );
 
 export default Homepage;
