+++ src/components/Logo.js
@@ -1,3 +1,6 @@
-const Logo = () => (<p className="logo">Math Magicians</p>);
+import logo from '../assets/images/logo.png';
+import styles from '../styles/Logo.module.css';
+
+const Logo = () => (<img src={logo} alt="Logo" className={styles.logo} />);
 
 export default Logo;
