+++ src/components/Navbar.js
@@ -1,0 +1,22 @@
+import { NavLink } from 'react-router-dom';
+import styles from '../styles/Navbar.module.css';
+
+const links = [
+  { path: '/', text: 'Home' },
+  { path: 'calculator', text: 'Calculator' },
+  { path: 'quote', text: 'Quote' },
+];
+
+const Navbar = () => (
+  <nav className={styles.navbar}>
+    <ul>
+      {links.map((link) => (
+        <li key={link.text}>
+          <NavLink to={link.path} className={styles.linkItem}>{link.text}</NavLink>
+        </li>
+      ))}
+    </ul>
+  </nav>
+);
+
+export default Navbar;
