+++ src/index.css
@@ -1,3 +1,7 @@
+html {
+  font-size: calc(15px + 0.390625vw);
+}
+
 body {
   margin: 0;
   font-family:
