+++ src/styles/Logo.module.css
@@ -1,0 +1,3 @@
+.logo {
+  width: 150px;
+}
