+++ src/styles/Navbar.module.css
@@ -1,0 +1,17 @@
+.navbar ul{
+  display: flex;
+  list-style: none;
+  flex-wrap: wrap;
+  gap: 1rem;
+}
+
+.navbar li {
+  font-size: 1rem;
+}
+
+.linkItem {
+  font-size: 1.115rem;
+  text-decoration: none;
+  color: #A437DB;
+  cursor: pointer;
+}
